ZTF File Source
